fix the bug, the user cannot delete other story and added the policy call UserStoryPolicy

+++ app/Http/Controllers/ImageController.php
@@ -37,6 +37,8 @@ class ImageController extends Controller
 
     public function deleteUserStory($profile,Image $image)
     {
+        $this->authorize('delete',$image);
+
         $image->delete();
         return back();
     }
